+++ js/packages/openinference-instrumentation-langchain/src/utils.ts
@@ -182,6 +182,9 @@ function getRoleFromMessageData(
   if (normalizedLangchainMessageClass.includes("function")) {
     return "function";
   }
+  if (normalizedLangchainMessageClass.includes("tool")) {
+    return "tool";
+  }
   if (
     normalizedLangchainMessageClass.includes("chat") &&
     isObject(messageData.kwargs) &&