+++ python/instrumentation/openinference-instrumentation-haystack/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.20](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-haystack-v0.1.19...python-openinference-instrumentation-haystack-v0.1.20) (2025-04-11)
+
+
+### Bug Fixes
+
+* increased minimum supported version of openinference-instrumentation to 0.1.27 ([#1507](https://github.com/Arize-ai/openinference/issues/1507)) ([a55edfa](https://github.com/Arize-ai/openinference/commit/a55edfa8900c1f36a73385c7d03f91cffadd85c4))
+
 ## [0.1.19](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-haystack-v0.1.18...python-openinference-instrumentation-haystack-v0.1.19) (2025-03-14)